renamed some files for more consistent naming scheme
