+++ software/control/core/__init__.py
@@ -5,8 +5,8 @@ os.environ["QT_API"] = "pyqt5"
 from .stream_handler import StreamHandler
 from .image_saver import ImageSaver
 from .configuration import Configuration, ConfigurationManager
-from .live_controller import LiveController
-from .navigation_controller import NavigationController
-from .slide_position_controller import SlidePositionController
+from .live import LiveController
+from .navigation import NavigationController
+from .slide_position import SlidePositionController
 from .autofocus import AutoFocusController
-from .multi_point_controller import MultiPointController
+from .multi_point import MultiPointController
